+++ internal/services/mocks.go
@@ -84,18 +84,13 @@ func NewRPCServiceMock(t interface {
 	return mock
 }
 
-type AccountSponsorshipServiceMock struct {
+type FeeBumpServiceMock struct {
 	mock.Mock
 }
 
-var _ AccountSponsorshipService = (*AccountSponsorshipServiceMock)(nil)
+var _ FeeBumpService = (*FeeBumpServiceMock)(nil)
 
-func (s *AccountSponsorshipServiceMock) SponsorAccountCreationTransaction(ctx context.Context, accountToSponsor string, signers []entities.Signer, assets []entities.Asset) (string, string, error) {
-	args := s.Called(ctx, accountToSponsor, signers, assets)
-	return args.String(0), args.String(1), args.Error(2)
-}
-
-func (s *AccountSponsorshipServiceMock) WrapTransaction(ctx context.Context, tx *txnbuild.Transaction) (string, string, error) {
+func (s *FeeBumpServiceMock) WrapTransaction(ctx context.Context, tx *txnbuild.Transaction) (string, string, error) {
 	args := s.Called(ctx, tx)
 	return args.String(0), args.String(1), args.Error(2)
 }
