+++ openapi/main.yaml
@@ -277,107 +277,6 @@ paths:
               example:
                 status: 500
                 error: An error occurred while processing this request.
-  /tx/create-sponsored-account:
-    post:
-      tags:
-        - Transactions
-      summary: Sponsor (pay the base reserve for) the creation of an account on the network
-      description: Create an account on the stellar network whose base reserve is paid for by the wallet backend's distribution account. Make sure that the total added weight of the partial signers is less than the total added weight of the full signers.
-      operationId: createSponsoredAccount
-      requestBody:
-        description: "Request body for configuring address signers with weights and types."
-        required: true
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                address:
-                  type: string
-                  description: "The address of the account to be created."
-                  example: "GBQ4HFNATBR4NXIDHDNSLAXZAWUFQ7DLMS5DUFWOBXNXFBOZFE3NQ243"
-                signers:
-                  type: array
-                  description: "List of signers with their respective addresses, types, and weights."
-                  items:
-                    type: object
-                    properties:
-                      address:
-                        type: string
-                        description: "The address of the signer."
-                        example: "GCXCF2T2FUY6UFYBLXNLJKGA4K6Y23KQEWTGV2X2YRBYOTP2RVYKY2DU"
-                      type:
-                        type: string
-                        description: "Type of the signer (full, partial)."
-                        example: "full"
-                      weight:
-                        type: integer
-                        description: "Weight of the signer."
-                        example: 20
-              required:
-                - address
-                - signers
-              example:
-                address: "GBQ4HFNATBR4NXIDHDNSLAXZAWUFQ7DLMS5DUFWOBXNXFBOZFE3NQ243"
-                signers:
-                  - address: "GCXCF2T2FUY6UFYBLXNLJKGA4K6Y23KQEWTGV2X2YRBYOTP2RVYKY2DU"
-                    type: "full"
-                    weight: 20
-                  - address: "GCAMJ67W77XEHUCPRTN3WGD6QG7XMJVWTU5KJMD4VDCI6FRUIO3X7B4T"
-                    type: "partial"
-                    weight: 10
-                  - address: "GC5E47IDYJAKVHXRLQM7Z2ABBARG6BAT6CYPEKTAZERS4H24KFUEKCA7"
-                    type: "partial"
-                    weight: 10
-      responses:
-        '200':
-          description: "Successful response containing network passphrase and the base64 transaction xdr string."
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  networkPassphrase:
-                    type: string
-                    description: "The passphrase of the network used for the transaction."
-                  transaction:
-                    type: string
-                    description: "The base64-encoded transaction envelope."
-              example:
-                networkPassphrase: "Test SDF Network ; September 2015"
-                transaction: "AAAAAgAAAADnY6MF7SMT2a2d6pt3i37Xx9IhaHqJ2lCcibNOISzhOwABX5AAFs2YAAAADAAAAAEAAAAAAAAAAAAAAABmYGswAAAAAAAAAAkAAAABAAAAAOdjowXtIxPZrZ3qm3eLftfH0iFoeonaUJyJs04hLOE7AAAAEAAAAACOC8v8STBDIULGM3FlZ6O7N3vHpNns7bcwRDFlIxTMiwAAAAEAAAAA52OjBe0jE9mtneqbd4t+18fSIWh6idpQnImzTiEs4TsAAAAAAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAAAAAAAAAAAAABAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAABQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAAApIt66UEG1IDSjJPfTQNheHz06pj5cjZjsCB3mMCRHhQAAABQAAAABAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAABQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAAD7c8Mp4XnAkJ0wzCcXHw7lxQNMgnG3c8auqJWB75WJywAAAAoAAAABAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAABQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAACIKVZwHT74BHJ6f1PVDd5HWMzEuPW1P5bw3dDipS9RzgAAAAoAAAABAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAABgAAAAFVU0RDAAAAAEI+fQXy7K+/7BkrIVo/G+lq7bjY5wJUq+NBPgIH3layf/////////8AAAABAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAABgAAAAFBUlNUAAAAAH8wYjTJienWf2nf2TEZi2APPWzmtkwiQHAftisIgyuHf/////////8AAAABAAAAAI4Ly/xJMEMhQsYzcWVno7s3e8ek2ezttzBEMWUjFMyLAAAAEQAAAAEAAAAAjgvL/EkwQyFCxjNxZWejuzd7x6TZ7O23MEQxZSMUzIsAAAAFAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAABAAAAFAAAAAEAAAAUAAAAAQAAABQAAAAAAAAAAAAAAAAAAAABISzhOwAAAEA2UXoQIAlm8nq2K8fniAbMdsBbWRdx6ay8p5dAJUGfqFGJ7rWf/pdZVQ8D3gcKXlBwTVhHDVCSBfj0UsTaHREO"
-        '400':
-          description: Bad Request
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  status:
-                    type: string
-                  error:
-                    type: string
-                    description: Details about the error
-                example:
-                  status: 400
-                  error: Validation Error.
-        '500':
-          description: Internal Server Error
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  status:
-                    type: string
-                  error:
-                    type: string
-                example:
-                  status: 500
-                  error: An error occurred while processing this request.
-              example:
-                status: 500
-                error: An error occurred while processing this request.
   /tx/create-fee-bump:
     post:
       tags:
